+++ src/compiler/compiler_internal.h
@@ -320,6 +320,7 @@ typedef struct
 	bool maydiscard : 1;
 	bool is_pure : 1;
 	bool noreturn : 1;
+	bool always_const : 1;
 } CalleeAttributes;
 
 typedef struct