+++ releasenotes.md
@@ -22,6 +22,7 @@
 - Fix incorrect override of optimization levels when using projects.
 - Add experimental `@noalias` attribute.
 - Add a `--run-once` option to delete the output file after running it.
+- Add `@const` attribute for macros, for better error messages with constant macros.
 
 ### Fixes